바킹독 4강 문제풀기

+++ 백준/바킹독/0x03_배열/10808_알파벳 개수.cpp
@@ -6,7 +6,7 @@
 /*   By: minskim2 <minskim2@student.42seoul.kr>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/03/22 18:03:33 by minskim2          #+#    #+#             */
-/*   Updated: 2022/03/22 19:36:55 by minskim2         ###   ########.fr       */
+/*   Updated: 2022/03/24 08:38:09 by minskim2         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,7 +19,7 @@ int	main(void) {
 	string word;
 	int alp[26] = {};
 	cin >> word;
-	for (int i=0; i<word.size(); i++)
+	for (size_t i=0; i<word.size(); i++)
 		alp[word[i]-'a']++;
 	for (int i:alp)
 		cout << i << " ";
